Implement SettingsScreen
Add Label gui component

Add screen resolutions

Still need to work on the actual implementation of the settings

+++ include/GameSave/GameSettings.h
@@ -16,7 +16,7 @@ namespace GameSave {
 
         int masterVolume = 128;
         int musicVolume = 128;
-        float sfxVolume = 128;
+        int sfxVolume = 128;
 
         json toJson() const {
             return {
